// src/chunithm/model/userGameOption.ts
export interface UserGameOptionItem {
  bgInfo: number;
  fieldColor: number;
  guideSound: number;
  soundEffect: number;
  guideLine: number;
  speed: number;
  optionSet: number;
  matching: number;
  judgePos: number;
  rating: number;
  judgeJustice: number;
  judgeAttack: number;
  headphone: number;
  playerLevel: number;
  successTap: number;
  successExTap: number;
  successSlideHold: number;
  successAir: number;
  successFlick: number;
  successSkill: number;
  successTapTimbre: number;
  privacy: number;
}